requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
sqlalchemy==1.4.46
flask-migrate==3.1.0
werkzeug==2.0.1
python-dotenv==0.19.0
flask-login==0.6.3
flask_wtf=1.2.2
wtforms=3.2.1
dnspython=2.7.0
email_validator=2.2.0